+++ cmds/enrollRewardProgram-testnet.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+set -e
+symbol-bootstrap enrollRewardProgram -t target/testnet-supernode  --useKnownRestGateways --password 1111 $1 $2 $3