+++ src/service/AgentCertificateService.ts
@@ -14,28 +14,70 @@
  * limitations under the License.
  */
 
+import { existsSync } from 'fs';
 import { join, resolve } from 'path';
+import { NetworkType } from 'symbol-sdk';
 import { LogType } from '../logger';
 import Logger from '../logger/Logger';
 import LoggerFactory from '../logger/LoggerFactory';
+import { CertificatePair } from '../model';
 import { BootstrapUtils } from './BootstrapUtils';
-
+import { CertificateService } from './CertificateService';
+import { CommandUtils } from './CommandUtils';
+import { KeyName } from './ConfigService';
 export interface AgentCertificateParams {
     readonly target: string;
     readonly user: string;
 }
 
+export interface AgentCertificates {
+    agent: CertificatePair;
+}
+
+export interface AgentCertificateMetadata {
+    readonly agentPublicKey: string;
+    readonly version: number;
+}
+
 const logger: Logger = LoggerFactory.getLogger(LogType.System);
 
 export class AgentCertificateService {
+    private static readonly METADATA_VERSION = 1;
     constructor(private readonly root: string, protected readonly params: AgentCertificateParams) {}
 
-    public async run(symbolServerToolsImage: string, name: string, customCertFolder?: string): Promise<void> {
+    public async run(
+        networkType: NetworkType,
+        symbolServerToolsImage: string,
+        name: string,
+        providedCertificates: AgentCertificates,
+        customCertFolder?: string,
+    ): Promise<void> {
         const copyFrom = `${this.root}/config/agent-cert`;
         const certFolder = customCertFolder || BootstrapUtils.getTargetNodesFolder(this.params.target, false, name, 'agent');
         await BootstrapUtils.mkdir(certFolder);
+
+        const metadataFile = join(certFolder, 'metadata.yml');
+        if (!(await this.shouldGenerateCertificate(metadataFile, providedCertificates))) {
+            logger.info(`Agent Certificates for node ${name} have been previously generated. Reusing...`);
+            return;
+        }
+        await BootstrapUtils.deleteFolder(certFolder);
+        await BootstrapUtils.mkdir(certFolder);
+        const newCertsFolder = join(certFolder, 'new_certs');
+        await BootstrapUtils.mkdir(newCertsFolder);
+
         const generatedContext = { name };
         await BootstrapUtils.generateConfiguration(generatedContext, copyFrom, certFolder, []);
+
+        const agentPrivateKey = await CommandUtils.resolvePrivateKey(
+            networkType,
+            providedCertificates.agent,
+            KeyName.Agent,
+            name,
+            'generating the Agent certificates',
+        );
+        BootstrapUtils.createDerFile(agentPrivateKey, join(certFolder, 'agent-ca.der'));
+
         const command = this.createCertCommands('/data');
         await BootstrapUtils.writeTextFile(join(certFolder, 'createAgentCertificate.sh'), command);
         const cmd = ['bash', 'createAgentCertificate.sh'];
@@ -53,20 +95,74 @@ export class AgentCertificateService {
             logger.error(BootstrapUtils.secureString(stderr));
             throw new Error('Certificate creation failed. Check the logs!');
         }
+
+        const certificates = CertificateService.getCertificates(stdout);
+        if (certificates.length != 1) {
+            throw new Error('Certificate creation failed. 1 certificates should have been created but got: ' + certificates.length);
+        }
+        const agentCertificate = certificates[0];
+        BootstrapUtils.validateIsTrue(agentCertificate.privateKey === agentPrivateKey, 'Invalid agent private key');
+        BootstrapUtils.validateIsTrue(agentCertificate.publicKey === providedCertificates.agent.publicKey, 'Invalid agent public key');
         logger.info(`Agent Certificate for node ${name} created`);
+        const metadata: AgentCertificateMetadata = {
+            version: AgentCertificateService.METADATA_VERSION,
+            agentPublicKey: providedCertificates.agent.publicKey,
+        };
+        await BootstrapUtils.writeYaml(metadataFile, metadata, undefined);
     }
+    private async shouldGenerateCertificate(metadataFile: string, providedCertificates: AgentCertificates): Promise<boolean> {
+        if (!existsSync(metadataFile)) {
+            return true;
+        }
+        try {
+            const metadata = BootstrapUtils.loadYaml(metadataFile, false) as AgentCertificateMetadata;
+            return (
+                metadata.agentPublicKey !== providedCertificates.agent.publicKey ||
+                metadata.version !== AgentCertificateService.METADATA_VERSION
+            );
+        } catch (e) {
+            logger.warn(`Cannot load agent certificate metadata from file ${metadataFile}. Error: ${e.message}`, e);
+            return true;
+        }
+    }
+
+    // Alternative reading the public key from agent-ca.pubkey.pem
+    // It's ok but it requires sshpk and unsure how to force an regeneration
+    // private async shouldGenerateCertificate(certFolder: string, providedCertificates: AgentCertificates): Promise<boolean> {
+    //   const pemFile = join(certFolder, 'agent-ca.pubkey.pem');
+    //   if (!existsSync(pemFile)) {
+    //     return true;
+    //   }
+    //   try {
+    //     const data = readFileSync(pemFile);
+    //     const key = sshpk.parseKey(data, 'pem');
+    //     const storedPublicKey = Convert.uint8ToHex((key.part as any).A.data);
+    //     return storedPublicKey !== providedCertificates.agent.publicKey;
+    //   } catch (e) {
+    //     logger.warn(`Cannot extract public key from file ${pemFile}. Error: ${e.message}`, e);
+    //     return true;
+    //   }
+    // }
 
     private createCertCommands(target: string): string {
         return `set -e
-        cd ${target}
+cd ${target}
+
+# prepare dirs/files
+chmod 700 new_certs
+touch index.txt
+
 # Creating a self signed certificate for the agent. This will be created by bootstrap when setting up a supernode
-openssl genrsa -out agent-key.pem 4096
-openssl req -new -config agent.cnf -key agent-key.pem -out agent-csr.pem
-openssl x509 -req -extfile agent.cnf -days 999 -in agent-csr.pem -out agent-crt.pem -signkey agent-key.pem
+
+cat agent-ca.der | openssl pkey -inform DER -outform PEM -out agent-ca.key.pem
+openssl pkey -inform pem -in agent-ca.key.pem -text -noout
+openssl pkey -in agent-ca.key.pem -pubout -out agent-ca.pubkey.pem
+
+# create agent CA CSR
+openssl req -config agent-ca.cnf -key agent-ca.key.pem -new -out agent-ca.csr.pem
 
 rm createAgentCertificate.sh
-rm agent-csr.pem
-rm agent.cnf
+rm agent-ca.der
 
 echo "Certificate Created"
 `;